schedule: specify meeting that is multiple days away
Currently, the only dates permitted by this action are 'today' and 'tomorrow'.
This prevents scenarios in which, for example, someone would like to create the
issue on Monday and record ideas/topics throughout the week for a meeting on
Friday.

Therefore, this change introduces the concept of 'x days from now'. In addition
to today and tomorrow, users can now specify, for example, 4 days from now for
the scenario described above.